Initial commits
I believe these are initial commits. Also, something interesting to note: Scrapy saves its cache file in the same location that the script is run. In this case, that means about 25 MB of data. I do not know if that is a lot, but I may need to change it if GitHub won't let me upload it.